Move the root layout's chrome into src/components/layout/
`Nav`, `NavLink` and `Footer` are the three components `src/app/layout.tsx`
renders on every route, and the only ones with that property. Grouping
them says which components a change reaches all five pages through.

Nav's two `ui/` imports reach up a level; nothing else in the six files
changes. `NavLink` is only ever rendered by `Nav`, so it moves as part of
the same subject rather than as a shared shape.
